jpa spring-data

+++ jpa/chap06-spring-data-jpa-lecture-source/src/main/java/com/ohgiraffers/springdatajpa/menu/repository/CategoryRepository.java
@@ -1,6 +1,5 @@
 package com.ohgiraffers.springdatajpa.menu.repository;
 
-import com.ohgiraffers.springdatajpa.menu.dto.CategoryDTO;
 import com.ohgiraffers.springdatajpa.menu.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -8,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 
 
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 
     // JPQL or Native Query를 작성하는 방법
     @Query(
